Refactor route directory structure
