+++ routes/toplevel/home.js
@@ -1,30 +1,12 @@
 'use strict';
 const express = require('express');
 const router = express.Router();
-const config = require('config');
 const rp = require('request-promise');
-const grants = require('../bin/data/grantnav.json');
-const logger = require('../logger');
+const grants = require('../../bin/data/grantnav.json');
+const logger = require('../../logger');
 
 router.get('/', (req, res, next) => {
     res.render('pages/index', {});
-});
-
-// handle contrast shifter
-router.get('/contrast/:mode', (req, res, next) => {
-    res.cacheControl = { maxAge: 1 };
-    let duration = 6 * 30 * 24 * 60 * 60; // 6 months
-    let cookieName = config.get('contrastCookie.name');
-    let redirectUrl = req.query.url || '/';
-    if (req.params.mode === 'high') {
-        res.cookie(cookieName, req.params.mode, {
-            maxAge: duration,
-            httpOnly: false
-        });
-    } else {
-        res.clearCookie(cookieName);
-    }
-    res.redirect(redirectUrl);
 });
 
 router.get('/lookup', (req, res, next) => {
